requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai>=1.58.0
python-dotenv==1.0.1
